etl_service/requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
Deprecated==1.2.13
packaging==21.3
psycopg2-binary==2.9.4
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
elastic-transport==8.4.0
elasticsearch[async]==7.9.1
